+++ engine/src/main/java/dev/gamekit/ui/widgets/Slider.java
@@ -7,7 +7,7 @@ import dev.gamekit.ui.events.ChangeEvent;
 import dev.gamekit.ui.events.MouseEvent;
 import dev.gamekit.utils.Bounds;
 import dev.gamekit.utils.Constraints;
-import dev.gamekit.utils.EngineImage;
+import dev.gamekit.utils.Picture;
 import dev.gamekit.utils.Position;
 
 import java.awt.*;
@@ -17,11 +17,11 @@ import static dev.gamekit.utils.Math.clamp;
 /** A {@link Progress} widget extension which adjusts a value by moving a slider */
 @WidgetBuilder
 public class Slider extends Progress implements MouseEvent.Handler {
-  public static final EngineImage DEFAULT_THUMB_BG =
+  public static final Picture DEFAULT_THUMB_BG =
     IO.getImageSliceWithInsets("default-sprites.png", 470, 346, 32, 32, 8, 8, 8, 8);
 
   @WidgetBuilderField(fallback = "dev.gamekit.ui.widgets.Slider.DEFAULT_THUMB_BG")
-  public EngineImage thumbBackground;
+  public Picture thumbBackground;
   @WidgetBuilderField(fallback = "32")
   public Integer thumbWidth;
   @WidgetBuilderField(fallback = "32")