+++ engine/src/main/java/dev/gamekit/ui/widgets/Spinner.java
@@ -5,7 +5,7 @@ import dev.gamekit.annotations.WidgetBuilder;
 import dev.gamekit.core.IO;
 import dev.gamekit.settings.ImageInterpolation;
 import dev.gamekit.utils.Constraints;
-import dev.gamekit.utils.EngineImage;
+import dev.gamekit.utils.Picture;
 
 import java.awt.*;
 import java.awt.image.BufferedImage;
@@ -13,7 +13,7 @@ import java.awt.image.BufferedImage;
 /** A {@link Leaf} which renders a loading indicator */
 @WidgetBuilder
 public class Spinner extends Leaf {
-  private static final EngineImage LOADING_BG = IO.getImage("default-loader.png");
+  private static final Picture LOADING_BG = IO.getImage("default-loader.png");
   private static final int BG_SLICE_WIDTH = 200;
   private static final int BG_SLICE_HEIGHT = 200;
   private static final ImageInterpolation INTERPOLATION = ImageInterpolation.BICUBIC;