+++ samples/src/main/java/Demo6EntityComponents.java
@@ -46,7 +46,7 @@ public class Demo6EntityComponents extends Scene {
 
   @Override
   protected void start() {
-    RigidBody.DEBUG_DRAW = true;
+    RigidBody.DEBUG = true;
     Collider.DEBUG = true;
 
     addChild(new BoxFrame());